fix differential output
splitlines() splits on newline character but also removes the char from
the end of the line. Since unified_diff() expects the end of line
character to be a part of each line, the diff output is somewhat
mangled.

Keep the end-of-line character on the actual content and stop adding it
to all lines further down.

Signed-off-by: Liam Beguin <lvb@xiphos.com>